Assignment4/requirements.txt
numpy<2
torch
torchvision
jupyter
matplotlib
scipy
scikit-learn
opencv-python
pillow